Replace composition panel with KAB per AB#336514 instead of showing both
