+++ src/frontend/src/containers/tables/population/HouseholdCompositionTable/index.ts
@@ -1,2 +1,1 @@
 export { HouseholdCompositionTable } from './HouseholdCompositionTable';
-export { KnownAffectedBeneficiariesTable } from './KnownAffectedBeneficiariesTable';
